Sketching out some initial ideas.
Nothing is solid for now, but my current plan is to allow a class to
include `::Equalish::Models::Canonical`, which will add some
methods that the class can use to define how it can be compared to other
classes that are equal to it.  I haven't thought this out 100% yet, but
I think it's a good starting point.

At the most basic, something like:
```
class OfficialModel
  canonicalize some_field, [:alt_field_name_1, :alt_field_name_2] do
    ['a', 'b', 'c']
  end
end
```

will allow objects which:
```
include ::Equalish::Models::Equivocal

canonical_model OfficialModel
```
to test for equality via `OfficialModel`.